Implement search state management, stdin reader, and worker thread for engine operations
- Added `SearchState` enum to manage the search states (Idle, Searching, StopRequested).
- Implemented `spawn_stdin_reader` function to handle input commands from stdin and communicate via channels.
- Created `WorkerMessage` enum for communication between worker threads and the main thread, including messages for search results, errors, and engine return.
- Developed `EngineReturnGuard` to ensure safe engine management and automatic return on drop.
- Implemented `search_worker` function to execute the search logic, handle callbacks, and manage session updates.
- Added `wait_for_worker_with_timeout` function to manage worker thread completion and handle timeouts gracefully.
