+++ packages/rust-core/crates/engine-cli/src/main.rs
@@ -1,170 +1,37 @@
 // USI (Universal Shogi Interface) adapter
 
+mod command_handler;
 mod engine_adapter;
+mod helpers;
 mod search_session;
+mod state;
+mod stdin_reader;
 mod usi;
 mod utils;
+mod worker;
 
-use anyhow::{anyhow, Result};
+use anyhow::Result;
 use clap::Parser;
+use command_handler::{handle_command, CommandContext};
 use crossbeam_channel::{bounded, select, unbounded, Receiver, Sender};
-use engine_adapter::{EngineAdapter, EngineError};
-use engine_core::engine::controller::Engine;
-use engine_core::search::SearchLimits;
+use engine_adapter::EngineAdapter;
 use engine_core::usi::move_to_usi;
+use helpers::{generate_fallback_move, send_bestmove_once};
 use search_session::SearchSession;
-use std::io::{self, BufRead};
+use state::SearchState;
 use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::{Arc, Mutex, MutexGuard};
+use std::sync::{Arc, Mutex};
 use std::thread::{self, JoinHandle};
-use std::time::{Duration, Instant};
-use usi::output::{Score, SearchInfo};
+use std::time::Duration;
+use stdin_reader::spawn_stdin_reader;
+use usi::output::SearchInfo;
 use usi::{
-    ensure_flush_on_exit, flush_final, parse_usi_command, send_info_string, send_response,
-    send_response_or_exit, UsiCommand, UsiResponse,
+    ensure_flush_on_exit, flush_final, send_info_string, send_response, UsiCommand, UsiResponse,
 };
-use utils::lock_or_recover_generic;
+use worker::{lock_or_recover_adapter, WorkerMessage};
 
 // Constants for timeout and channel management
-const MIN_JOIN_TIMEOUT: Duration = Duration::from_secs(5);
 const CHANNEL_SIZE: usize = 1024;
-const SELECT_TIMEOUT: Duration = Duration::from_millis(50);
-
-/// Specialized lock_or_recover for EngineAdapter with state reset
-fn lock_or_recover_adapter(mutex: &Mutex<EngineAdapter>) -> MutexGuard<'_, EngineAdapter> {
-    match mutex.lock() {
-        Ok(guard) => guard,
-        Err(poisoned) => {
-            log::error!("EngineAdapter mutex was poisoned, attempting recovery with state reset");
-            let mut guard = poisoned.into_inner();
-
-            // Force reset engine state to safe defaults
-            guard.force_reset_state();
-
-            // Try to notify GUI about the reset
-            let _ = send_info_string(
-                "Engine state reset due to error recovery. Please send 'isready' to reinitialize.",
-            );
-
-            guard
-        }
-    }
-}
-
-/// Perform fallback move generation with graduated strategy
-///
-/// This function attempts to generate a move using increasingly simple methods:
-/// 1. Use partial result from interrupted search (instant)
-/// 2. Run quick shallow search (depth 3, ~10-100ms)
-/// 3. Generate emergency move using heuristics only (~1ms)
-///
-/// All operations are synchronous but designed to be fast.
-/// Total worst-case time: ~100ms (dominated by quick_search)
-fn generate_fallback_move(
-    engine: &Arc<Mutex<EngineAdapter>>,
-    partial_result: Option<(String, u32, i32)>,
-    allow_null_move: bool,
-) -> Result<String> {
-    // Stage 1: Use partial result if available (instant)
-    if let Some((best_move, depth, score)) = partial_result {
-        // Validate the partial result move before using it
-        let adapter = lock_or_recover_adapter(engine);
-        if adapter.is_legal_move(&best_move) {
-            log::info!(
-                "Using validated partial result: move={best_move}, depth={depth}, score={score}"
-            );
-            return Ok(best_move);
-        } else {
-            log::warn!("Partial result move {best_move} is illegal, proceeding to Stage 2");
-        }
-    }
-
-    // Stage 2: Try quick shallow search (depth 3, typically 10-50ms, max 100ms)
-    log::info!("Attempting quick shallow search");
-    let shallow_result = {
-        let mut engine = lock_or_recover_adapter(engine);
-        match engine.quick_search() {
-            Ok(move_str) => {
-                log::info!("Quick search successful: {move_str}");
-                Some(move_str)
-            }
-            Err(e) => {
-                log::warn!("Quick search failed: {e}");
-                None
-            }
-        }
-    };
-
-    if let Some(move_str) = shallow_result {
-        return Ok(move_str);
-    }
-
-    // Stage 3: Try emergency move generation (heuristic only, ~1ms)
-    log::info!("Attempting emergency move generation");
-    let emergency_result = {
-        let engine = lock_or_recover_adapter(engine);
-        engine.generate_emergency_move()
-    };
-
-    match emergency_result {
-        Ok(move_str) => {
-            log::info!("Generated emergency move: {move_str}");
-            Ok(move_str)
-        }
-        Err(EngineError::NoLegalMoves) => {
-            log::error!("No legal moves available - position is checkmate or stalemate");
-            Ok("resign".to_string())
-        }
-        Err(EngineError::EngineNotAvailable(msg)) if msg.contains("Position not set") => {
-            if allow_null_move {
-                log::error!("Position not set - returning null move (0000)");
-                // Return null move (0000) which most GUIs handle gracefully
-                // Note: This is not defined in USI spec but widely supported
-                Ok("0000".to_string())
-            } else {
-                log::error!("Position not set - returning resign");
-                Ok("resign".to_string())
-            }
-        }
-        Err(e) => {
-            log::error!("Failed to generate fallback move: {e}");
-            if allow_null_move {
-                // Return null move for better GUI compatibility
-                // Note: This is not defined in USI spec but widely supported
-                Ok("0000".to_string())
-            } else {
-                // Return resign as per USI spec
-                Ok("resign".to_string())
-            }
-        }
-    }
-}
-
-/// Unified function to send bestmove response and update state atomically
-fn send_bestmove_once(
-    best_move: String,
-    ponder: Option<String>,
-    search_state: &mut SearchState,
-    bestmove_sent: &mut bool,
-) -> Result<()> {
-    // Check if already sent
-    if *bestmove_sent {
-        log::debug!("Bestmove already sent, ignoring duplicate: {best_move}");
-        return Ok(());
-    }
-
-    // Log before sending (while we still own the values)
-    log::info!("Bestmove sent: {best_move}, ponder: {ponder:?}");
-
-    // Send the response
-    send_response(UsiResponse::BestMove { best_move, ponder })?;
-
-    // Update state atomically
-    *search_state = SearchState::Idle;
-    *bestmove_sent = true;
-
-    Ok(())
-}
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
@@ -177,412 +44,6 @@ struct Args {
     /// Note: null move is not defined in USI spec but handled by most GUIs
     #[arg(long)]
     allow_null_move: bool,
-}
-
-/// Guard to ensure engine is returned on drop (for panic safety)
-struct EngineReturnGuard {
-    engine: Option<Engine>,
-    tx: Sender<WorkerMessage>,
-    search_id: u64,
-}
-
-impl EngineReturnGuard {
-    fn new(engine: Engine, tx: Sender<WorkerMessage>, search_id: u64) -> Self {
-        Self {
-            engine: Some(engine),
-            tx,
-            search_id,
-        }
-    }
-}
-
-impl std::ops::Deref for EngineReturnGuard {
-    type Target = Engine;
-
-    fn deref(&self) -> &Self::Target {
-        self.engine.as_ref().expect("Engine already taken")
-    }
-}
-
-impl std::ops::DerefMut for EngineReturnGuard {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.engine.as_mut().expect("Engine already taken")
-    }
-}
-
-impl Drop for EngineReturnGuard {
-    fn drop(&mut self) {
-        if let Some(engine) = self.engine.take() {
-            log::debug!("EngineReturnGuard: returning engine");
-
-            // Try to return engine through channel
-            match self.tx.try_send(WorkerMessage::EngineReturn(engine)) {
-                Ok(()) => {
-                    log::debug!("Engine returned successfully through channel");
-                }
-                Err(crossbeam_channel::TrySendError::Full(_)) => {
-                    // Channel is full - this shouldn't happen with unbounded channel
-                    log::error!("Channel full, cannot return engine");
-                    // Engine will be dropped here, which is safe
-                }
-                Err(crossbeam_channel::TrySendError::Disconnected(_)) => {
-                    // Channel is disconnected - main thread has exited
-                    log::warn!("Channel disconnected, cannot return engine");
-                    // Engine will be dropped here, which is safe
-                }
-            }
-
-            // Always try to send Finished message to signal completion (from guard)
-            let _ = self.tx.try_send(WorkerMessage::Finished {
-                from_guard: true,
-                search_id: self.search_id,
-            });
-        }
-    }
-}
-
-/// Search state management - tracks the current state of the search
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum SearchState {
-    /// No search is active
-    Idle,
-    /// Search is actively running
-    Searching,
-    /// Stop has been requested but search is still running
-    StopRequested,
-}
-
-impl SearchState {
-    /// Check if we're in any searching state
-    fn is_searching(&self) -> bool {
-        matches!(self, SearchState::Searching | SearchState::StopRequested)
-    }
-
-    /// Check if we can start a new search
-    fn can_start_search(&self) -> bool {
-        matches!(self, SearchState::Idle)
-    }
-
-    /// Check if we should accept a bestmove
-    fn can_accept_bestmove(&self) -> bool {
-        matches!(self, SearchState::Searching | SearchState::StopRequested)
-    }
-}
-
-/// Messages from worker thread to main thread
-enum WorkerMessage {
-    Info(SearchInfo),
-
-    /// Iteration completed with committed results
-    IterationComplete {
-        session: Box<SearchSession>,
-        search_id: u64,
-    },
-
-    /// Search finished (use committed results from session)
-    SearchFinished {
-        session_id: u64,
-        root_hash: u64,
-        search_id: u64,
-    },
-
-    // Legacy messages (to be phased out)
-    BestMove {
-        best_move: String,
-        ponder_move: Option<String>,
-        search_id: u64,
-    },
-    /// Partial result available during search
-    PartialResult {
-        current_best: String,
-        depth: u32,
-        score: i32,
-        search_id: u64,
-    },
-    /// Thread finished - from_guard indicates if sent by EngineReturnGuard
-    Finished {
-        from_guard: bool,
-        search_id: u64,
-    },
-    Error {
-        message: String,
-        search_id: u64,
-    },
-    EngineReturn(Engine), // Return the engine after search
-}
-
-/// Context for handling USI commands
-struct CommandContext<'a> {
-    engine: &'a Arc<Mutex<EngineAdapter>>,
-    stop_flag: &'a Arc<AtomicBool>,
-    worker_tx: &'a Sender<WorkerMessage>,
-    worker_rx: &'a Receiver<WorkerMessage>,
-    worker_handle: &'a mut Option<JoinHandle<()>>,
-    search_state: &'a mut SearchState,
-    bestmove_sent: &'a mut bool,
-    current_search_timeout: &'a mut Duration,
-    search_id_counter: &'a mut u64,
-    current_search_id: &'a mut u64,
-    current_search_is_ponder: &'a mut bool,
-    current_session: &'a mut Option<SearchSession>,
-    allow_null_move: bool,
-}
-
-/// Calculate maximum expected search time from GoParams
-fn calculate_max_search_time(params: &usi::GoParams) -> Duration {
-    if params.infinite {
-        // For infinite search, use a large but reasonable timeout
-        return Duration::from_secs(3600); // 1 hour
-    }
-
-    if let Some(movetime) = params.movetime {
-        // Fixed time per move + margin
-        return Duration::from_millis(movetime + 1000);
-    }
-
-    // For time-based searches, estimate based on available time
-    let mut max_time = 0u64;
-
-    if let Some(wtime) = params.wtime {
-        max_time = max_time.max(wtime);
-    }
-    if let Some(btime) = params.btime {
-        max_time = max_time.max(btime);
-    }
-    if let Some(byoyomi) = params.byoyomi {
-        // Byoyomi could be used multiple times
-        let periods = params.periods.unwrap_or(1) as u64;
-        max_time = max_time.max(byoyomi * periods);
-    }
-
-    if max_time > 0 {
-        // Use half of available time + margin, with upper limit
-        // Cap at 10 seconds to avoid excessively long stop waits
-        Duration::from_millis((max_time / 2 + 2000).min(10000))
-    } else {
-        // Default timeout for depth/node limited searches
-        Duration::from_secs(10) // Reduced from 60s for better responsiveness
-    }
-}
-
-/// Wait for worker thread to finish with timeout
-fn wait_for_worker_with_timeout(
-    worker_handle: &mut Option<JoinHandle<()>>,
-    worker_rx: &Receiver<WorkerMessage>,
-    engine: &Arc<Mutex<EngineAdapter>>,
-    search_state: &mut SearchState,
-    timeout: Duration,
-) -> Result<()> {
-    let deadline = Instant::now() + timeout.max(MIN_JOIN_TIMEOUT);
-    let mut finished = false;
-    let mut engine_returned = false;
-    let mut finished_count = 0u32;
-
-    // Wait for Finished message AND EngineReturn message or timeout
-    loop {
-        select! {
-            recv(worker_rx) -> msg => {
-                match msg {
-                    Ok(WorkerMessage::Finished { from_guard, search_id: _ }) => {
-                        finished_count += 1;
-                        if !finished {
-                            log::debug!("Worker thread finished cleanly (from_guard: {from_guard})");
-                            finished = true;
-                            if engine_returned {
-                                break;
-                            }
-                        } else {
-                            log::trace!("Ignoring duplicate Finished message #{finished_count} (from_guard: {from_guard})");
-                        }
-                    }
-                    Ok(WorkerMessage::EngineReturn(returned_engine)) => {
-                        log::debug!("Engine returned from worker");
-                        let mut adapter = lock_or_recover_adapter(engine);
-                        adapter.return_engine(returned_engine);
-                        engine_returned = true;
-                        if finished {
-                            break;
-                        }
-                    }
-                    Ok(WorkerMessage::Info(info)) => {
-                        // Info messages during shutdown can be ignored
-                        log::trace!("Received info during shutdown: {info:?}");
-                    }
-                    Ok(WorkerMessage::BestMove { best_move, ponder_move, search_id }) => {
-                        // During shutdown, we may accept late bestmoves
-                        // For safety, we could check search_id here but during shutdown
-                        // we're more lenient since we're trying to clean up
-                        if search_state.can_accept_bestmove() {
-                            log::debug!("Accepting bestmove during shutdown (search_id: {search_id})");
-                            send_response_or_exit(UsiResponse::BestMove {
-                                best_move,
-                                ponder: ponder_move,
-                            });
-                            // Mark search as finished when bestmove is received
-                            *search_state = SearchState::Idle;
-                        } else {
-                            log::warn!("Ignoring late bestmove during shutdown: {best_move} (search_id: {search_id})");
-                        }
-                    }
-                    Ok(WorkerMessage::PartialResult { .. }) => {
-                        // Partial results during shutdown can be ignored
-                        log::trace!("PartialResult during shutdown - ignoring");
-                    }
-                    Ok(WorkerMessage::Error { message, search_id }) => {
-                        log::error!("Worker error during shutdown (search_id: {search_id}): {message}");
-                    }
-                    Ok(WorkerMessage::IterationComplete { .. }) => {
-                        // Iteration updates during shutdown can be ignored
-                        log::trace!("IterationComplete during shutdown - ignoring");
-                    }
-                    Ok(WorkerMessage::SearchFinished { .. }) => {
-                        // Search finished during shutdown can be ignored
-                        log::trace!("SearchFinished during shutdown - ignoring");
-                    }
-                    Err(_) => {
-                        log::error!("Worker channel closed unexpectedly");
-                        break;
-                    }
-                }
-            }
-            default(SELECT_TIMEOUT) => {
-                if Instant::now() > deadline {
-                    log::error!("Worker thread timeout after {:?}", timeout.max(MIN_JOIN_TIMEOUT));
-                    // Return error instead of exit for graceful handling
-                    return Err(anyhow::anyhow!("Worker thread timeout"));
-                }
-            }
-        }
-    }
-
-    // If we received Finished, join() should complete immediately
-    if finished {
-        if let Some(handle) = worker_handle.take() {
-            match handle.join() {
-                Ok(()) => log::debug!("Worker thread joined successfully"),
-                Err(_) => log::error!("Worker thread panicked"),
-            }
-        }
-    }
-
-    *search_state = SearchState::Idle;
-
-    // Drain any remaining messages in worker_rx
-    while let Ok(msg) = worker_rx.try_recv() {
-        match msg {
-            WorkerMessage::EngineReturn(returned_engine) => {
-                log::debug!("Engine returned during drain");
-                let mut adapter = lock_or_recover_adapter(engine);
-                adapter.return_engine(returned_engine);
-            }
-            _ => {
-                log::trace!("Drained message: {:?}", std::any::type_name_of_val(&msg));
-            }
-        }
-    }
-
-    Ok(())
-}
-
-/// Wait for any ongoing search to complete
-fn wait_for_search_completion(
-    search_state: &mut SearchState,
-    stop_flag: &Arc<AtomicBool>,
-    worker_handle: &mut Option<JoinHandle<()>>,
-    worker_rx: &Receiver<WorkerMessage>,
-    engine: &Arc<Mutex<EngineAdapter>>,
-) -> Result<()> {
-    if search_state.is_searching() {
-        *search_state = SearchState::StopRequested;
-        stop_flag.store(true, Ordering::Release);
-        wait_for_worker_with_timeout(
-            worker_handle,
-            worker_rx,
-            engine,
-            search_state,
-            MIN_JOIN_TIMEOUT,
-        )?;
-    }
-    Ok(())
-}
-
-/// Spawn stdin reader thread
-fn spawn_stdin_reader(cmd_tx: Sender<UsiCommand>) -> JoinHandle<()> {
-    thread::spawn(move || {
-        let stdin = io::stdin();
-        let reader = stdin.lock();
-
-        for line in reader.lines() {
-            match line {
-                Ok(line) => {
-                    let line = line.trim();
-                    if line.is_empty() {
-                        continue;
-                    }
-
-                    log::debug!("Received: {line}");
-
-                    match parse_usi_command(line) {
-                        Ok(cmd) => {
-                            // Use try_send to avoid blocking
-                            match cmd_tx.try_send(cmd) {
-                                Ok(()) => {}
-                                Err(crossbeam_channel::TrySendError::Full(_)) => {
-                                    log::warn!("Command channel full, dropping command");
-                                }
-                                Err(crossbeam_channel::TrySendError::Disconnected(_)) => {
-                                    log::debug!(
-                                        "Command channel disconnected, exiting stdin reader"
-                                    );
-                                    return;
-                                }
-                            }
-                        }
-                        Err(e) => {
-                            log::error!("Failed to parse command '{line}': {e}");
-                            // Invalid commands are silently ignored in USI protocol
-                        }
-                    }
-                }
-                Err(e) => {
-                    // Distinguish between EOF and actual errors
-                    match e.kind() {
-                        io::ErrorKind::UnexpectedEof | io::ErrorKind::BrokenPipe => {
-                            log::info!(
-                                "Stdin closed (EOF or broken pipe), shutting down gracefully"
-                            );
-                        }
-                        io::ErrorKind::Interrupted => {
-                            // EINTR - could retry, but for stdin it's safer to exit
-                            log::warn!("Stdin read interrupted, shutting down");
-                        }
-                        _ => {
-                            log::error!("Stdin read error: {e}");
-                        }
-                    }
-
-                    // Try to send quit command for graceful shutdown
-                    match cmd_tx.try_send(UsiCommand::Quit) {
-                        Ok(()) => {
-                            log::debug!("Sent quit command for graceful shutdown");
-                        }
-                        Err(_) => {
-                            log::debug!("Failed to send quit command, channel likely closed");
-                        }
-                    }
-                    break;
-                }
-            }
-        }
-
-        // Reached here = normal EOF. GUI closed the pipe, send quit
-        match cmd_tx.try_send(UsiCommand::Quit) {
-            Ok(()) => log::info!("Sent quit command after EOF"),
-            Err(_) => log::debug!("Channel closed before quit after EOF"),
-        }
-
-        log::debug!("Stdin reader thread exiting (EOF)");
-    })
 }
 
 fn main() {
@@ -631,7 +92,7 @@ fn run_engine(allow_null_move: bool) -> Result<()> {
     let mut worker_handle: Option<JoinHandle<()>> = None;
     let mut search_state = SearchState::Idle;
     let mut bestmove_sent = false; // Track if bestmove has been sent for current search
-    let mut current_search_timeout = MIN_JOIN_TIMEOUT;
+    let mut current_search_timeout = helpers::MIN_JOIN_TIMEOUT;
     let mut search_id_counter = 0u64;
     let mut current_search_id = 0u64;
     let mut current_search_is_ponder = false; // Track if current search is ponder
@@ -892,12 +353,12 @@ fn run_engine(allow_null_move: bool) -> Result<()> {
     // Stop any ongoing search with timeout
     stop_flag.store(true, Ordering::Release);
     if search_state.is_searching() {
-        wait_for_worker_with_timeout(
+        worker::wait_for_worker_with_timeout(
             &mut worker_handle,
             &worker_rx,
             &engine,
             &mut search_state,
-            MIN_JOIN_TIMEOUT,
+            helpers::MIN_JOIN_TIMEOUT,
         )?;
     }
 
@@ -915,955 +376,6 @@ fn run_engine(allow_null_move: bool) -> Result<()> {
 
     log::debug!("Shutdown complete");
     Ok(())
-}
-
-fn handle_command(command: UsiCommand, ctx: &mut CommandContext) -> Result<()> {
-    match command {
-        UsiCommand::Usi => {
-            send_response(UsiResponse::IdName("RustShogi 1.0".to_string()))?;
-            send_response(UsiResponse::IdAuthor("RustShogi Team".to_string()))?;
-
-            // Send available options
-            {
-                let engine = lock_or_recover_adapter(ctx.engine);
-                for option in engine.get_options() {
-                    send_response(UsiResponse::Option(option.to_string()))?;
-                }
-            }
-
-            send_response(UsiResponse::UsiOk)?;
-        }
-
-        UsiCommand::IsReady => {
-            // Initialize engine if needed
-            {
-                let mut engine = lock_or_recover_adapter(ctx.engine);
-                engine.initialize()?;
-            }
-            send_response(UsiResponse::ReadyOk)?;
-        }
-
-        UsiCommand::Position {
-            startpos,
-            sfen,
-            moves,
-        } => {
-            log::info!(
-                "Handling position command - startpos: {startpos}, sfen: {sfen:?}, moves: {moves:?}"
-            );
-            // Wait for any ongoing search to complete before updating position
-            wait_for_search_completion(
-                ctx.search_state,
-                ctx.stop_flag,
-                ctx.worker_handle,
-                ctx.worker_rx,
-                ctx.engine,
-            )?;
-
-            let mut engine = lock_or_recover_adapter(ctx.engine);
-            engine.set_position(startpos, sfen.as_deref(), &moves)?;
-            log::info!("Position command completed");
-        }
-
-        UsiCommand::Go(params) => {
-            log::info!("Received go command with params: {params:?}");
-
-            // Stop any ongoing search and ensure engine is available
-            wait_for_search_completion(
-                ctx.search_state,
-                ctx.stop_flag,
-                ctx.worker_handle,
-                ctx.worker_rx,
-                ctx.engine,
-            )?;
-
-            // Add a small delay to ensure clean state transition
-            thread::sleep(Duration::from_millis(10));
-
-            // Reset stop flag and bestmove_sent flag
-            ctx.stop_flag.store(false, Ordering::Release);
-            *ctx.bestmove_sent = false; // Reset for new search
-
-            // Verify we can start a new search (defensive check)
-            if !ctx.search_state.can_start_search() {
-                log::error!("Cannot start search in state: {:?}", ctx.search_state);
-                return Err(anyhow!("Invalid state for starting search"));
-            }
-
-            // Verify position is set before starting search
-            {
-                let engine = lock_or_recover_adapter(ctx.engine);
-                if !engine.has_position() {
-                    log::error!("Cannot start search: position not set");
-                    send_response(UsiResponse::BestMove {
-                        best_move: "resign".to_string(),
-                        ponder: None,
-                    })?;
-                    return Ok(());
-                }
-            }
-
-            // Increment search ID for new search
-            *ctx.search_id_counter += 1;
-            *ctx.current_search_id = *ctx.search_id_counter;
-            let search_id = *ctx.current_search_id;
-            log::info!("Starting new search with ID: {search_id}, ponder: {}", params.ponder);
-
-            // Calculate timeout for this search
-            *ctx.current_search_timeout = calculate_max_search_time(&params);
-
-            // Track if this is a ponder search
-            *ctx.current_search_is_ponder = params.ponder;
-
-            // Clone necessary data for worker thread
-            let engine_clone = Arc::clone(ctx.engine);
-            let stop_clone = Arc::clone(ctx.stop_flag);
-            let tx_clone = ctx.worker_tx.clone();
-
-            // Spawn worker thread for search with panic safety
-            let handle = thread::spawn(move || {
-                log::debug!("Worker thread spawned");
-                let result = std::panic::catch_unwind(|| {
-                    search_worker(engine_clone, params, stop_clone, tx_clone.clone(), search_id);
-                });
-
-                if let Err(e) = result {
-                    log::error!("Worker thread panicked: {e:?}");
-                    // Send error message to main thread
-                    let _ = tx_clone.send(WorkerMessage::Error {
-                        message: "Worker thread panicked".to_string(),
-                        search_id,
-                    });
-                    let _ = tx_clone.send(WorkerMessage::Finished {
-                        from_guard: false,
-                        search_id,
-                    });
-                }
-            });
-
-            *ctx.worker_handle = Some(handle);
-            *ctx.search_state = SearchState::Searching;
-            log::info!("Worker thread handle stored, search_state = Searching");
-
-            // Don't block - return immediately
-        }
-
-        UsiCommand::Stop => {
-            log::info!("Received stop command, search_state = {:?}", *ctx.search_state);
-            log::debug!("Stop command received, entering stop handler");
-
-            // Early return if bestmove already sent or not searching
-            if *ctx.bestmove_sent || !ctx.search_state.is_searching() {
-                log::debug!("Stop while idle or already sent bestmove -> ignore");
-                return Ok(());
-            }
-
-            // Signal stop to worker thread
-            if ctx.search_state.is_searching() {
-                *ctx.search_state = SearchState::StopRequested;
-                ctx.stop_flag.store(true, Ordering::Release);
-                log::info!("Stop flag set to true, search_state = StopRequested");
-
-                // Debug: Verify stop flag was actually set
-                let stop_value = ctx.stop_flag.load(Ordering::Acquire);
-                log::info!("Stop flag verification: {stop_value}");
-
-                // First try to use committed best from session immediately
-                if let Some(ref session) = *ctx.current_session {
-                    let adapter = lock_or_recover_adapter(ctx.engine);
-                    if let Some(position) = adapter.get_position() {
-                        if let Ok((best_move, ponder)) =
-                            adapter.validate_and_get_bestmove(session, position)
-                        {
-                            // Send info string about bestmove source
-                            let depth =
-                                session.committed_best.as_ref().map(|b| b.depth).unwrap_or(0);
-                            let score_str = session
-                                .committed_best
-                                .as_ref()
-                                .map(|b| match &b.score {
-                                    search_session::Score::Cp(cp) => format!("cp {cp}"),
-                                    search_session::Score::Mate(mate) => format!("mate {mate}"),
-                                })
-                                .unwrap_or_else(|| "unknown".to_string());
-                            send_info_string(format!(
-                                "bestmove_from=session_on_stop depth={depth} score={score_str}"
-                            ))?;
-
-                            log::info!(
-                                "Sending committed bestmove from session on stop: {best_move}"
-                            );
-                            return send_bestmove_once(
-                                best_move,
-                                ponder,
-                                ctx.search_state,
-                                ctx.bestmove_sent,
-                            );
-                        }
-                    }
-                }
-
-                // Get overhead from adapter to determine timeouts
-                let overhead_ms = {
-                    let adapter = lock_or_recover_adapter(ctx.engine);
-                    adapter.get_last_overhead_ms()
-                };
-
-                // Use longer timeouts for byoyomi mode
-                let stage1_timeout = if overhead_ms >= usi::DEFAULT_BYOYOMI_OVERHEAD_MS {
-                    Duration::from_millis(500) // Byoyomi mode: wait longer for in-flight messages
-                } else {
-                    Duration::from_millis(100) // Normal mode: quick wait
-                };
-                let total_timeout = if overhead_ms >= usi::DEFAULT_BYOYOMI_OVERHEAD_MS {
-                    Duration::from_millis(1000) // Byoyomi mode: up to 1 second total
-                } else {
-                    Duration::from_millis(150) // Normal mode: quick fallback
-                };
-
-                // Wait for bestmove with staged timeouts
-                let start = Instant::now();
-                let mut partial_result: Option<(String, u32, i32)> = None;
-                let mut stage = 1;
-
-                loop {
-                    let elapsed = start.elapsed();
-
-                    // Stage transition logic
-                    if stage == 1 && elapsed > stage1_timeout {
-                        stage = 2;
-                        log::debug!(
-                            "Stop handler stage 2: trying fallback after {}ms",
-                            elapsed.as_millis()
-                        );
-                    }
-
-                    if elapsed > total_timeout {
-                        // Timeout - use fallback strategy
-                        log::warn!("Timeout waiting for bestmove after stop command");
-                        // Log timeout error
-                        log::debug!("Stop command timeout: {:?}", EngineError::Timeout);
-
-                        if *ctx.current_search_is_ponder {
-                            // Ponder search - don't send bestmove (USI protocol)
-                            log::info!(
-                                "Ponder search timeout, not sending bestmove (USI protocol)"
-                            );
-                            *ctx.search_state = SearchState::Idle;
-                            *ctx.current_search_is_ponder = false; // Reset ponder flag
-                            break;
-                        }
-
-                        // Use emergency fallback (session already tried at the beginning)
-                        match generate_fallback_move(
-                            ctx.engine,
-                            partial_result.clone(),
-                            ctx.allow_null_move,
-                        ) {
-                            Ok(move_str) => {
-                                // Send info string about fallback source
-                                if let Some((_, depth, score)) = partial_result {
-                                    send_info_string(format!(
-                                        "bestmove_from=partial_result depth={depth} score={score}"
-                                    ))?;
-                                } else {
-                                    send_info_string("bestmove_from=emergency_fallback_timeout")?;
-                                }
-                                log::debug!("Sending emergency fallback bestmove: {move_str}");
-                                send_bestmove_once(
-                                    move_str,
-                                    None,
-                                    ctx.search_state,
-                                    ctx.bestmove_sent,
-                                )?;
-                            }
-                            Err(e) => {
-                                log::error!("Emergency fallback move generation failed: {e}");
-                                send_bestmove_once(
-                                    "resign".to_string(),
-                                    None,
-                                    ctx.search_state,
-                                    ctx.bestmove_sent,
-                                )?;
-                            }
-                        }
-                        break;
-                    }
-
-                    // Check for bestmove message
-                    match ctx.worker_rx.try_recv() {
-                        Ok(WorkerMessage::BestMove {
-                            best_move,
-                            ponder_move,
-                            search_id,
-                        }) => {
-                            // Only accept if it's for current search and not pondering
-                            if search_id == *ctx.current_search_id {
-                                if !*ctx.current_search_is_ponder {
-                                    send_bestmove_once(
-                                        best_move,
-                                        ponder_move,
-                                        ctx.search_state,
-                                        ctx.bestmove_sent,
-                                    )?;
-                                    *ctx.current_search_is_ponder = false; // Reset ponder flag after sending bestmove
-                                } else {
-                                    // Ponder search stopped - don't send bestmove
-                                    log::debug!("Ponder search stopped, not sending bestmove");
-                                    *ctx.search_state = SearchState::Idle;
-                                    *ctx.current_search_is_ponder = false; // Reset ponder flag
-                                }
-                                break;
-                            }
-                        }
-                        Ok(WorkerMessage::Info(info)) => {
-                            // Forward info messages
-                            let _ = send_response(UsiResponse::Info(info));
-                        }
-                        Ok(WorkerMessage::PartialResult {
-                            current_best,
-                            depth,
-                            score,
-                            search_id,
-                        }) => {
-                            // Store partial result for fallback only if it's from current search
-                            if search_id == *ctx.current_search_id {
-                                partial_result = Some((current_best, depth, score));
-                            }
-                        }
-                        Ok(WorkerMessage::IterationComplete { session, search_id }) => {
-                            // Update current session
-                            if search_id == *ctx.current_search_id {
-                                *ctx.current_session = Some(*session);
-                            }
-                        }
-                        Ok(WorkerMessage::SearchFinished {
-                            session_id: _,
-                            root_hash: _,
-                            search_id,
-                        }) => {
-                            // Handle SearchFinished in stop command context
-                            if search_id == *ctx.current_search_id && !*ctx.current_search_is_ponder
-                            {
-                                log::info!(
-                                    "SearchFinished received in stop handler, sending bestmove"
-                                );
-                                // Try to use session-based bestmove
-                                if let Some(ref session) = *ctx.current_session {
-                                    let adapter = lock_or_recover_adapter(ctx.engine);
-                                    if let Some(position) = adapter.get_position() {
-                                        match adapter.validate_and_get_bestmove(session, position) {
-                                            Ok((best_move, ponder)) => {
-                                                // Send info string about bestmove source
-                                                let depth = session
-                                                    .committed_best
-                                                    .as_ref()
-                                                    .map(|b| b.depth)
-                                                    .unwrap_or(0);
-                                                let score_str = session
-                                                    .committed_best
-                                                    .as_ref()
-                                                    .map(|b| match &b.score {
-                                                        search_session::Score::Cp(cp) => {
-                                                            format!("cp {cp}")
-                                                        }
-                                                        search_session::Score::Mate(mate) => {
-                                                            format!("mate {mate}")
-                                                        }
-                                                    })
-                                                    .unwrap_or_else(|| "unknown".to_string());
-                                                send_info_string(format!("bestmove_from=session_in_stop_handler depth={depth} score={score_str}"))?;
-
-                                                log::info!("Sending bestmove from stop handler: {best_move}");
-                                                send_bestmove_once(
-                                                    best_move,
-                                                    ponder,
-                                                    ctx.search_state,
-                                                    ctx.bestmove_sent,
-                                                )?;
-                                                break;
-                                            }
-                                            Err(e) => {
-                                                log::warn!("Session validation failed in stop handler: {e}");
-                                                // Continue to wait for BestMove or use fallback
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        Ok(WorkerMessage::Finished {
-                            from_guard,
-                            search_id,
-                        }) => {
-                            // Only process if it's for current search
-                            if search_id == *ctx.current_search_id {
-                                if *ctx.current_search_is_ponder {
-                                    // Ponder search - don't send bestmove (USI protocol)
-                                    log::debug!("Ponder search finished without bestmove, not sending fallback (USI protocol)");
-                                    *ctx.search_state = SearchState::Idle;
-                                    *ctx.current_search_is_ponder = false; // Reset ponder flag
-                                    break;
-                                }
-
-                                // Normal search - use fallback strategy
-                                log::warn!(
-                                    "Worker finished without bestmove (from_guard: {from_guard})"
-                                );
-                                match generate_fallback_move(
-                                    ctx.engine,
-                                    partial_result.clone(),
-                                    ctx.allow_null_move,
-                                ) {
-                                    Ok(move_str) => {
-                                        // Send info string about fallback source
-                                        if let Some((_, depth, score)) = partial_result {
-                                            send_info_string(format!("bestmove_from=partial_result_on_finish depth={depth} score={score}"))?;
-                                        } else {
-                                            send_info_string(
-                                                "bestmove_from=emergency_fallback_on_finish",
-                                            )?;
-                                        }
-                                        send_bestmove_once(
-                                            move_str,
-                                            None,
-                                            ctx.search_state,
-                                            ctx.bestmove_sent,
-                                        )?;
-                                    }
-                                    Err(e) => {
-                                        log::error!("Fallback move generation failed: {e}");
-                                        send_bestmove_once(
-                                            "resign".to_string(),
-                                            None,
-                                            ctx.search_state,
-                                            ctx.bestmove_sent,
-                                        )?;
-                                    }
-                                }
-                                break;
-                            }
-                        }
-                        _ => {
-                            thread::sleep(Duration::from_millis(10));
-                        }
-                    }
-                }
-            }
-        }
-
-        UsiCommand::PonderHit => {
-            // Handle ponder hit
-            let mut engine = lock_or_recover_adapter(ctx.engine);
-            // Mark that we're no longer in pure ponder mode
-            *ctx.current_search_is_ponder = false;
-            match engine.ponder_hit() {
-                Ok(()) => log::debug!("Ponder hit successfully processed"),
-                Err(e) => log::debug!("Ponder hit ignored: {e}"),
-            }
-        }
-
-        UsiCommand::SetOption { name, value } => {
-            let mut engine = lock_or_recover_adapter(ctx.engine);
-            engine.set_option(&name, value.as_deref())?;
-        }
-
-        UsiCommand::GameOver { result } => {
-            // Stop any ongoing search
-            ctx.stop_flag.store(true, Ordering::Release);
-
-            // Notify engine of game result
-            let mut engine = lock_or_recover_adapter(ctx.engine);
-            engine.game_over(result);
-        }
-
-        UsiCommand::UsiNewGame => {
-            // ShogiGUI extension - new game notification
-            // Stop any ongoing search
-            wait_for_search_completion(
-                ctx.search_state,
-                ctx.stop_flag,
-                ctx.worker_handle,
-                ctx.worker_rx,
-                ctx.engine,
-            )?;
-
-            // Reset engine state for new game
-            let mut engine = lock_or_recover_adapter(ctx.engine);
-            engine.new_game();
-            log::debug!("New game started");
-        }
-
-        UsiCommand::Quit => {
-            // Quit is handled in main loop
-            unreachable!("Quit should be handled in main loop");
-        }
-    }
-
-    Ok(())
-}
-
-/// Worker thread function for search
-fn search_worker(
-    engine_adapter: Arc<Mutex<EngineAdapter>>,
-    params: usi::GoParams,
-    stop_flag: Arc<AtomicBool>,
-    tx: Sender<WorkerMessage>,
-    search_id: u64,
-) {
-    log::debug!("Search worker thread started with params: {params:?}");
-
-    // Set up info callback with partial result tracking
-    let tx_info = tx.clone();
-    let tx_partial = tx.clone();
-    let last_partial_depth = Arc::new(Mutex::new(0u32));
-    let info_callback = move |info: SearchInfo| {
-        // Always send the info message
-        let _ = tx_info.send(WorkerMessage::Info(info.clone()));
-
-        // Send partial result at certain depth intervals
-        if let (Some(depth), Some(score), Some(pv)) =
-            (info.depth, info.score.as_ref(), info.pv.first())
-        {
-            // Check if we should send a partial result
-            let should_send = {
-                let mut last_depth = lock_or_recover_generic(&last_partial_depth);
-                if depth >= *last_depth + 5 || (depth >= 10 && depth > *last_depth) {
-                    *last_depth = depth;
-                    true
-                } else {
-                    false
-                }
-            };
-
-            if should_send {
-                let score_value = match score {
-                    Score::Cp(cp) => *cp,
-                    Score::Mate(mate) => {
-                        // Convert mate score to centipawn equivalent
-                        if *mate > 0 {
-                            30000 - (*mate * 100)
-                        } else {
-                            -30000 - (*mate * 100)
-                        }
-                    }
-                };
-
-                log::debug!(
-                    "Sending partial result: move={pv}, depth={depth}, score={score_value}"
-                );
-                let _ = tx_partial.send(WorkerMessage::PartialResult {
-                    current_best: pv.clone(),
-                    depth,
-                    score: score_value,
-                    search_id,
-                });
-            }
-        }
-    };
-
-    // Take engine out and prepare search
-    let was_ponder = params.ponder;
-    log::debug!("Attempting to take engine from adapter");
-    let (engine, position, limits, ponder_hit_flag) = {
-        let mut adapter = lock_or_recover_adapter(&engine_adapter);
-        log::debug!("Adapter lock acquired, calling take_engine");
-        match adapter.take_engine() {
-            Ok(engine) => {
-                log::debug!("Engine taken successfully, preparing search");
-                match adapter.prepare_search(&params, stop_flag.clone()) {
-                    Ok((pos, lim, flag)) => {
-                        log::debug!("Search prepared successfully");
-                        (engine, pos, lim, flag)
-                    }
-                    Err(e) => {
-                        // Return engine and send error
-                        adapter.return_engine(engine);
-                        log::error!("Search preparation error: {e}");
-                        let _ = tx.send(WorkerMessage::Error {
-                            message: e.to_string(),
-                            search_id,
-                        });
-
-                        // Try to generate emergency move before resigning (only if not pondering)
-                        if !params.ponder {
-                            match adapter.generate_emergency_move() {
-                                Ok(emergency_move) => {
-                                    log::info!(
-                                        "Generated emergency move after preparation error: {emergency_move}"
-                                    );
-                                    if let Err(e) = tx.send(WorkerMessage::BestMove {
-                                        best_move: emergency_move,
-                                        ponder_move: None,
-                                        search_id,
-                                    }) {
-                                        log::error!("Failed to send emergency move: {e}");
-                                    }
-                                }
-                                Err(_) => {
-                                    // Only resign if no legal moves available
-                                    if let Err(e) = tx.send(WorkerMessage::BestMove {
-                                        best_move: "resign".to_string(),
-                                        ponder_move: None,
-                                        search_id,
-                                    }) {
-                                        log::error!(
-                                            "Failed to send resign after preparation error: {e}"
-                                        );
-                                    }
-                                }
-                            }
-                        } else {
-                            log::info!(
-                                "Ponder preparation error, not sending bestmove (USI protocol)"
-                            );
-                        }
-
-                        let _ = tx.send(WorkerMessage::Finished {
-                            from_guard: false,
-                            search_id,
-                        });
-                        return;
-                    }
-                }
-            }
-            Err(e) => {
-                log::error!("Failed to take engine: {e}");
-                let _ = tx.send(WorkerMessage::Error {
-                    message: e.to_string(),
-                    search_id,
-                });
-
-                // Try to generate emergency move from adapter (only if not pondering)
-                if !params.ponder {
-                    log::info!("Attempting to generate emergency move after engine take failure");
-                    match adapter.generate_emergency_move() {
-                        Ok(emergency_move) => {
-                            log::info!(
-                                "Generated emergency move after engine take error: {emergency_move}"
-                            );
-                            if let Err(e) = tx.send(WorkerMessage::BestMove {
-                                best_move: emergency_move,
-                                ponder_move: None,
-                                search_id,
-                            }) {
-                                log::error!("Failed to send emergency move: {e}");
-                            }
-                        }
-                        Err(_) => {
-                            // Only resign if no legal moves available
-                            if let Err(e) = tx.send(WorkerMessage::BestMove {
-                                best_move: "resign".to_string(),
-                                ponder_move: None,
-                                search_id,
-                            }) {
-                                log::error!("Failed to send resign after engine take error: {e}");
-                            }
-                        }
-                    }
-                } else {
-                    log::info!("Ponder engine take error, not sending bestmove (USI protocol)");
-                }
-
-                let _ = tx.send(WorkerMessage::Finished {
-                    from_guard: false,
-                    search_id,
-                });
-                return;
-            }
-        }
-    }; // Lock released here
-
-    // Keep ponder_hit_flag for checking later
-    let ponder_hit_flag_ref = ponder_hit_flag.clone();
-
-    // Update limits with ponder_hit_flag if present
-    let limits = if let Some(ref flag) = ponder_hit_flag {
-        SearchLimits {
-            ponder_hit_flag: Some(flag.clone()),
-            ..limits
-        }
-    } else {
-        limits
-    };
-
-    // Explicitly drop ponder_hit_flag (it's used internally by the engine)
-    drop(ponder_hit_flag);
-
-    // Create search session
-    let mut session = SearchSession::new(search_id, position.hash);
-
-    // Wrap session in Arc<Mutex<>> for shared access in callback
-    let session_arc = Arc::new(Mutex::new(session.clone()));
-    let session_for_callback = session_arc.clone();
-    let tx_for_iteration = tx.clone();
-
-    // Track last committed depth to avoid committing same depth multiple times
-    let last_committed_depth = Arc::new(Mutex::new(0u32));
-    let last_committed_depth_cb = last_committed_depth.clone();
-
-    // Create info callback that updates session
-    let enhanced_info_callback = move |info: SearchInfo| {
-        // Call original callback
-        info_callback(info.clone());
-
-        // Update session on each iteration completion
-        if !info.pv.is_empty() {
-            // Only commit when depth increases (not for same depth updates)
-            let depth = info.depth.unwrap_or(0);
-            {
-                let mut last = lock_or_recover_generic(&last_committed_depth_cb);
-                if depth <= *last {
-                    // Same or lower depth - skip commit
-                    return;
-                }
-                *last = depth;
-            }
-            // Note: The info.pv contains USI strings which lose piece type information.
-            // This is a limitation of the current architecture where SearchInfo uses strings.
-            // For now, we parse them back but accept the loss of piece type info.
-            // TODO: Consider passing original Move objects through a different channel
-            // to preserve full move information including piece types.
-            if !info.pv.is_empty() {
-                // Extract raw score from Score enum
-                let raw_score = match info.score {
-                    Some(Score::Cp(cp)) => cp,
-                    Some(Score::Mate(mate)) => {
-                        // Convert mate score to raw score format
-                        // Positive mate = we're winning
-                        let mate_score =
-                            engine_core::search::constants::MATE_SCORE - (mate.abs() * 2);
-                        if mate > 0 {
-                            mate_score
-                        } else {
-                            -mate_score
-                        }
-                    }
-                    None => 0,
-                };
-
-                // Convert PV from USI strings to Move objects
-                // WARNING: This loses piece type information from the original moves
-                let pv_moves: Vec<_> = info
-                    .pv
-                    .iter()
-                    .filter_map(|m| engine_core::usi::parse_usi_move(m).ok())
-                    .collect();
-
-                // Update session
-                if let Ok(mut session_guard) = session_for_callback.lock() {
-                    session_guard.update_current_best(info.depth.unwrap_or(0), raw_score, pv_moves);
-                    session_guard.commit_iteration();
-
-                    // Send IterationComplete message
-                    let _ = tx_for_iteration.send(WorkerMessage::IterationComplete {
-                        session: Box::new(session_guard.clone()),
-                        search_id,
-                    });
-                }
-            }
-        }
-    };
-
-    // Wrap engine in guard for panic safety
-    let mut engine_guard = EngineReturnGuard::new(engine, tx.clone(), search_id);
-
-    // Execute search without holding the lock
-    log::info!("Calling execute_search_static");
-    let result = EngineAdapter::execute_search_static(
-        &mut engine_guard,
-        position.clone(),
-        limits,
-        Box::new(enhanced_info_callback),
-    );
-    log::info!("execute_search_static returned: {:?}", result.is_ok());
-
-    // Handle result
-    match result {
-        Ok(extended_result) => {
-            // Update session with final result using the original Move objects
-            log::debug!(
-                "Worker: Updating session with best_move: {}, ponder_move: {:?}, depth: {}",
-                extended_result.best_move,
-                extended_result.ponder_move,
-                extended_result.depth
-            );
-
-            // Use actual search result data with original Move objects
-            // This preserves the piece type information that would be lost if we
-            // converted to USI strings and back
-            session.update_current_best(
-                extended_result.depth,
-                extended_result.score,
-                extended_result.pv, // Original Move objects with piece types preserved
-            );
-            session.commit_iteration();
-            // Clean up ponder state if needed
-            {
-                let mut adapter = lock_or_recover_adapter(&engine_adapter);
-                adapter.cleanup_after_search(was_ponder);
-            }
-
-            // Check if ponderhit occurred during ponder search
-            let ponder_hit_occurred = if was_ponder {
-                // Check if ponder_hit_flag was set during search
-                ponder_hit_flag_ref
-                    .as_ref()
-                    .map(|flag| flag.load(Ordering::Acquire))
-                    .unwrap_or(false)
-            } else {
-                false
-            };
-
-            // Send best move if:
-            // - Not a ponder search OR
-            // - Ponder search that was converted via ponderhit
-            if !was_ponder || ponder_hit_occurred {
-                log::info!(
-                    "Sending search completion: was_ponder={was_ponder}, ponder_hit={ponder_hit_occurred}"
-                );
-                // Send completed session instead of raw bestmove
-                if let Err(e) = tx.send(WorkerMessage::IterationComplete {
-                    session: Box::new(session.clone()),
-                    search_id,
-                }) {
-                    log::error!("Failed to send iteration complete: {e}");
-                }
-                // Also send SearchFinished to indicate we're done
-                if let Err(e) = tx.send(WorkerMessage::SearchFinished {
-                    session_id: session.id,
-                    root_hash: session.root_hash,
-                    search_id,
-                }) {
-                    log::error!("Failed to send search finished: {e}");
-                }
-            } else {
-                log::info!("Ponder search without ponderhit, not sending bestmove (USI protocol)");
-            }
-        }
-        Err(e) => {
-            log::error!("Search error: {e}");
-            // Engine will be returned automatically by EngineReturnGuard::drop
-
-            // Clean up ponder state if needed
-            {
-                let mut adapter = lock_or_recover_adapter(&engine_adapter);
-                adapter.cleanup_after_search(was_ponder);
-            }
-
-            // Try to generate emergency move before sending error
-            let emergency_result = {
-                let adapter = lock_or_recover_adapter(&engine_adapter);
-                adapter.generate_emergency_move()
-            };
-
-            if stop_flag.load(Ordering::Acquire) {
-                // Check if ponderhit occurred for ponder search
-                let ponder_hit_occurred = if was_ponder {
-                    ponder_hit_flag_ref
-                        .as_ref()
-                        .map(|flag| flag.load(Ordering::Acquire))
-                        .unwrap_or(false)
-                } else {
-                    false
-                };
-
-                // Stopped by user - send bestmove if:
-                // - Not a ponder search OR
-                // - Ponder search that was converted via ponderhit
-                if !was_ponder || ponder_hit_occurred {
-                    // Normal search or ponder-hit search that was stopped - send emergency move
-                    match emergency_result {
-                        Ok(emergency_move) => {
-                            log::info!("Generated emergency move after stop: {emergency_move}");
-                            if let Err(e) = tx.send(WorkerMessage::BestMove {
-                                best_move: emergency_move,
-                                ponder_move: None,
-                                search_id,
-                            }) {
-                                log::error!("Failed to send emergency move: {e}");
-                            }
-                        }
-                        Err(_) => {
-                            // Only resign if no legal moves
-                            if let Err(e) = tx.send(WorkerMessage::BestMove {
-                                best_move: "resign".to_string(),
-                                ponder_move: None,
-                                search_id,
-                            }) {
-                                log::error!("Failed to send resign after stop: {e}");
-                            }
-                        }
-                    }
-                } else {
-                    // Ponder search that was stopped (not ponderhit) - don't send bestmove
-                    log::info!("Ponder search stopped, not sending bestmove (USI protocol)");
-                }
-            } else {
-                // Other error - send error and try emergency move
-                // Check if ponderhit occurred for ponder search
-                let ponder_hit_occurred = if was_ponder {
-                    ponder_hit_flag_ref
-                        .as_ref()
-                        .map(|flag| flag.load(Ordering::Acquire))
-                        .unwrap_or(false)
-                } else {
-                    false
-                };
-
-                let _ = tx.send(WorkerMessage::Error {
-                    message: e.to_string(),
-                    search_id,
-                });
-
-                // Send bestmove if not ponder OR ponder was converted via ponderhit
-                if !was_ponder || ponder_hit_occurred {
-                    match emergency_result {
-                        Ok(emergency_move) => {
-                            log::info!(
-                                "Generated emergency move after search error: {emergency_move}"
-                            );
-                            if let Err(e) = tx.send(WorkerMessage::BestMove {
-                                best_move: emergency_move,
-                                ponder_move: None,
-                                search_id,
-                            }) {
-                                log::error!("Failed to send emergency move: {e}");
-                            }
-                        }
-                        Err(_) => {
-                            // Only resign if no legal moves
-                            if let Err(e) = tx.send(WorkerMessage::BestMove {
-                                best_move: "resign".to_string(),
-                                ponder_move: None,
-                                search_id,
-                            }) {
-                                log::error!("Failed to send resign after error: {e}");
-                            }
-                        }
-                    }
-                } else {
-                    log::info!(
-                        "Ponder search error without ponderhit, not sending bestmove (USI protocol)"
-                    );
-                }
-            }
-        }
-    }
-
-    // Always send Finished at the end - use blocking send to ensure delivery
-    match tx.send(WorkerMessage::Finished {
-        from_guard: false,
-        search_id,
-    }) {
-        Ok(()) => log::debug!("Finished message sent successfully"),
-        Err(e) => {
-            log::error!("Failed to send Finished message: {e}. Channel might be closed.");
-            // This is a critical error but we can't do much about it
-        }
-    }
-
-    log::debug!("Search worker finished");
 }
 
 #[cfg(test)]
